+++ migrations/20260125154714-create-countries-table.js
@@ -44,12 +44,6 @@ module.exports = {
           comment:
             'ISO 3166-1 alpha-2 two-letter country codes and ISO 3166-1 alpha-3 three-letter country codes of the country.',
         },
-        surface_area: {
-          type: Sequelize.STRING(15),
-          allowNull: true,
-          defaultValue: null,
-          comment: 'Approximate surface area of the country (measured in km²).',
-        },
         tld: {
           type: Sequelize.STRING(10),
           allowNull: false,