+++ migrations/20260125161137-create-sub-divisions-table.js
@@ -4,7 +4,7 @@
 module.exports = {
   async up(queryInterface, Sequelize) {
     await queryInterface.createTable(
-      'geo_sub_divisions',
+      'geo_divisions',
       {
         id: {
           type: Sequelize.INTEGER,
@@ -42,18 +42,18 @@ module.exports = {
       }
     );
 
-    await queryInterface.addIndex('geo_sub_divisions', ['country_id'], {
+    await queryInterface.addIndex('geo_divisions', ['country_id'], {
       name: 'country',
     });
 
-    await queryInterface.addIndex('geo_sub_divisions', ['capital_id'], {
+    await queryInterface.addIndex('geo_divisions', ['capital_id'], {
       name: 'capital',
     });
 
-    await queryInterface.addConstraint('geo_sub_divisions', {
+    await queryInterface.addConstraint('geo_divisions', {
       fields: ['country_id'],
       type: 'foreign key',
-      name: 'geo_sub_divisions_ibfk_1',
+      name: 'geo_divisions_ibfk_1',
       references: {
         table: 'geo_countries',
         field: 'id',
@@ -62,10 +62,10 @@ module.exports = {
       onUpdate: 'CASCADE',
     });
 
-    await queryInterface.addConstraint('geo_sub_divisions', {
+    await queryInterface.addConstraint('geo_divisions', {
       fields: ['capital_id'],
       type: 'foreign key',
-      name: 'geo_sub_divisions_ibfk_2',
+      name: 'geo_divisions_ibfk_2',
       references: {
         table: 'geo_cities',
         field: 'id',
@@ -76,6 +76,6 @@ module.exports = {
   },
 
   async down(queryInterface, _Sequelize) {
-    await queryInterface.dropTable('geo_sub_divisions');
+    await queryInterface.dropTable('geo_divisions');
   },
 };