+++ seeders/12-base-countries-languages.seed.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+  get: () => {
+    const countriesLanguages = [];
+
+    return countriesLanguages;
+  },
+};